requirements.txt
streamlit==1.28.1
websocket-client==1.6.4
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
ccxt==4.1.77
asyncio-mqtt==0.16.1
aiohttp==3.9.1
